Support components that render no output or text output
Support components returning nullish values (null, boolean) that render
no output or strings, which render as text nodes.

In the process support was also added for passing these values to
`render`, since the values that can be passed to `render` are the same
that can be returned from custom components.