Do not set `dom` field for custom component `Component` node
This paves the way to supporting components that return arrays, since
such a component could produce more than a single top-level DOM element.

This change adds overhead to various operations because the top-level
DOM nodes can no longer just be read off a single field. It might be
necessary to add a separate field in future to track the root nodes
produced by a component tree.