Implement modern context API, except for `Context.Consumer`
Implement the `createContext` and `useContext` functions. The
`Context.Consumer` API is not currently implemented. It isn't needed for
basic context usage which only requires the `Context.Provider` class and
`useContext` hook.

This involved a refactoring of how a `HookState` instance interacts with
its associated component to make it easier to add new interactions in
future.