Switch package to use ES modules everywhere
Use native ES modules everywhere for consistency between source and
tests and to allow the output of TypeScript compilation to function both
as input to Rollup packaging and execution by mocha.

Node ES module support requires that file extensions are always
explicitly specified. TypeScript does not generate the ".js" extension
itself but counter-intuitively does allow it to be specified in the
original source [1].

[1] https://github.com/microsoft/TypeScript/issues/42151#issuecomment-753506506