Fix non-working event phase tests
These tests "passed" previously but an uncaught exception was thrown
afterwards. Their logic was broken since in order to differentiate
between bubble-phase and capture-phase events the listener needs to be
attached to a parent of the element that the event is dispatched to.
Event handlers on the element itself will always see the `eventPhase`
property `Event.AT_TARGET`.