Add an `ErrorBoundary` component to handle errors during rendering
The official React APIs require the use of a class component to add an
error boundary. Since ureact doesn't have class components, we need an
alternative. There has been discussion about providing a hook to handle
errors but [1] and [2] explain why hooks are not a good fit for this use
case.

The approach taken here is to add a builtin `ErrorBoundary` component
with a `handler` prop which is invoked when an error occurs in the
subtree below the `ErrorBoundary`.

[1] https://github.com/facebook/react/issues/14347#issuecomment-456861860
[2] https://overreacted.io/why-isnt-x-a-hook/