Allow setting `className` and other SVG props
Always set attributes rather than properties for SVG elements. The
immediate reason this doesn't work for many properties is that SVG
properties such as `width`, `height` and `className` are read-only.

However attributes can be set. In the case of the `className` prop, it
needs to be mapped to the `class` attr.

We could potentially implement a more generic solution which tests
whether a given property is read-only for an element and fall back to
setting an attribute in that case. The current implementation is
consistent with Preact however.