Change DOM property setting approach
Change the approach to setting DOM properties to avoid needing to
special-case SVGs and also avoiding recalculating certain data each time
a property is set.

Instead compute metadata about DOM properties the first time that a
particular prop is set on an element of a given type and use this to
determine which logic to execute for a prop. Part of this metadata is
whether an Element has a writable/settable property of a given name. If
it does then setting the prop writes to the property, otherwise an
attribute is set instead.