Automatically use pixel values if property supports it
For consistency with React/Preact, automatically convert numeric style
values to pixel units. Unlike these other implementations, attempt to
dynamically detect whether the property supports pixel units rather than
using a hardcoded list of unitless properties [1][2]. The detection
approach is to try to simply write a pixel-valued property to a test
element and mark the property as supporting or non-supporting pixel
valued properties depending on whether that succeeds or fails.

[1] https://reactjs.org/docs/dom-elements.html#style
[2] https://github.com/facebook/react/blob/95feb0e701a5ae20996e8cc6c4acd0f504d5985a/packages/react-dom/src/shared/CSSProperty.js#L11