Fix a bug causing incorrect DOM ordering after state update
The logic to compute the location to insert a component's output at after a state
update was incorrect. If none of the component's previous siblings
produced DOM output then a loop would traverse up the tree to check the
next level up. However after moving up a level we need to change which
component we look for in the parent's output to get the current
component's sibling index.