Fix bug with flushing updates scheduled during another update flush
If a component scheduled an update during a flush of earlier updates,
`_scheduleUpdates` would check `this._pendingUpdates` and see it was
non-empty and therefore assume that a flush had already been scheduled.
As a result the update scheduled during the initial flush would never
get flushed itself.

Fix the issue by making `_flushUpdates` loop as long as there are
pending updates and ensure that the set of pending updates is empty when
it returns.