Refactor DOM prop setting to more explicitly represent property type
I'm not sure if this will end up being the best representation, but it
feels cleaner to me to explicitly represent the type of DOM property
(property, attribute, event etc.) as a field of `PropMeta`. This also
allows reducing the number of fields.

This refactoring highlighted that unsetting the
`dangerouslySetInnerHTML` property was not accounted for in
`unsetProperty`.