Add a partial implementation of the Enzyme testing API
Add a partial implementation of the Enzyme API ([1]) for testing components.
This provides a convenient way to render a component in a test and query
/ traverse the rendered tree.

Rather than implement an Enzyme adapter, I want to see if instead a new
implementation of a compatible API works better instead.

The initial implementation includes the subset of the `mount` API that:

 1. Relevant for components built using the subset of React APIs
    supported by ureact (function components, hooks, modern context)

 2. We have actually used at Hypothesis

Changes in this commit:

 - Add initial implementation of Enzyme testing API subset in
   `src/enzyme.ts` and tests in `test/enzyme.js`
 - Modify `src/render.ts` to expose the rendered component tree and the
   `forEachDomRoot` method which is useful to traverse it.
 - Modify packaging configuration to generate separate bundles for the
   main library functionality (`dist/index.js`), the Enzyme testing API
   (`dist/enzyme.js`) and additional test utilities
   (`dist/test-utils.js`)

[1] https://github.com/enzymejs/enzyme/blob/master/docs/api/mount.md