Split `Component` into two separate interfaces
Split `Component` into two interfaces. One, `BaseComponent`, contains the essential
rendered structure of the tree that describes the output seen by the
user with the addition of custom components. The other extends this
essential structure with the data relating to internal state of the
component and precomputed pointers/flags used internally during
rendering.

The `BaseComponent` can be exposed for use by other modules without
exposing them to implementation details of the renderer.