Fix updating of `ref` when DOM component is replaced during a render
When a DOM component is replaced with a different one during a render
due to a change in `key` or `type`, the ref was not updated correctly if
shared between the old and new component. It would end up as `null`
instead of being set to the new DOM node. This happened because the old
component was unmounted after the new one was created and when
unmounting, the `ref` was unconditionally cleared.