Reduce duplicated logic amongst different task queues
Replace the separate fields for different task queues in `src/render.js`
with an array of queues where the index is a task type. The task type is
specified by a `TASK_` constant in `hooks.js`.

This allows simplifying some code which needs to do the same thing with
each of the three queues.