Handle errors during effects and effect cleanup
Handle errors when running effects and effect cleanup in the same way as
errors during rendering. If the component containing the effect has an
`ErrorBoundary` ancestor, that boundary's `handler` prop is invoked with
the error. Otherwise the entire component tree is unmounted.

This behavior is consistent with how React handles errors during
effects.