Switch `activeRoots` to be a WeakMap
Change `activeRoots` from a map to a `WeakMap` to avoid holding onto
references to roots once their container elements are removed from the
DOM.

The `act` testing utility needs to be able to enumerate all roots that
became "dirty" during the `act` call, and WeakMaps are not enumerable,
so a separate internal `dirtyRoots` set has been added for this purpose.
A side effect of this change is that `act` will only flush roots that
became dirty during the `act` callback.