Document which Enzyme methods are not yet implemented
Add a comment for each method listed in the Enzyme API docs [1] which is
not currently implemented. Some of these will never be implemented
because they do not apply to components implemented with hooks. Others
are simply not implemented yet because they haven't been used in the
test suites which the adapter has been tested against.

[1] https://github.com/enzymejs/enzyme/blob/master/docs/api/mount.md